Mnt/drop py2734

See merge request fsl/fslpy!75

+++ tests/test_imagewrapper.py
@@ -465,7 +465,10 @@ def _test_expansion(coverage, slices, volumes, expansions):
     # coverage, or in one of the expansions.
     dimranges = []
     for d in range(ndims):
-        dimranges.append(np.linspace(nc[0, d], nc[1, d], nc[1, d] / 5, dtype=np.uint32))
+        dimranges.append(np.linspace(nc[0, d],
+                                     nc[1, d],
+                                     int(nc[1, d] / 5),
+                                     dtype=np.uint32))
 
     points = it.product(*dimranges)
